+++ apps/prune/src/bin/prune.ts
@@ -2,7 +2,7 @@
  * An executable file that prune backuped files in a storage service.
  * Execute with --help to see usage instructions.
  */
-import { PruneCommand } from '@awesome-backup/core';
+import { PruneCommand } from '@awesome-backup/commands';
 import { PACKAGE_VERSION } from '../config/version';
 
 const pruneCommand = new PruneCommand();
