+++ packages/commands/__tests__/commands/list.test.ts
