+++ packages/commands/jest.config.ts
@@ -82,9 +82,6 @@ export default {
 
   // A map from regular expressions to module names or to arrays of module names that allow to stub out resources with a single module
   // moduleNameMapper: {},
-  moduleNameMapper: {
-    "^@awesome-backup/core(.*)": "<rootDir>/src/$1"
-  },
 
   // An array of regexp pattern strings, matched against all module paths before considered 'visible' to the module loader
   // modulePathIgnorePatterns: [],
