+++ packages/commands/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@awesome-backup/core",
+  "name": "@awesome-backup/commands",
   "version": "0.0.1",
-  "description": "core module of awesome-backup",
+  "description": "commands of awesome-backup",
   "author": "Ryu Sato <ryu@weseek.co.jp>",
   "license": "MIT",
   "homepage": "https://github.com/ryu-sato/awesome-backup#readme",
@@ -28,6 +28,7 @@
     "lint:fix": "eslint src --ext .ts --fix"
   },
   "dependencies": {
+    "@awesome-backup/storage-service-clients": "^0.0.1",
     "@aws-sdk/client-s3": "^3.45.0",
     "@google-cloud/storage": "^5.16.1",
     "axios": "^0.24.0",
