+++ packages/commands/src/commands/list.ts
