+++ packages/commands/src/commands/prune.ts
