+++ packages/commands/src/config/logger/config.dev.ts
