+++ packages/commands/src/config/logger/config.prod.ts
