--- packages/core/src/index.ts
@@ -1,10 +1,0 @@
-export * from './commands/backup';
-export * from './commands/list';
-export * from './commands/prune';
-export * from './commands/restore';
-export * from './storage-service-clients/factory';
-export * from './storage-service-clients/s3';
-export * from './storage-service-clients/s3-config';
-export * from './storage-service-clients/gcs';
-export * from './storage-service-clients/types';
-export * from './utils/tar';
