+++ packages/storage-service-clients/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@awesome-backup/mongodb-restore",
+  "name": "@awesome-backup/storage-service-clients",
   "version": "0.0.1",
-  "description": "Restore MongoDB database from backuped file on Amazon S3 or Google Cloud Storage. You can set a custom S3 endpoint to use S3 based services like DigitalOcean Spaces instead of Amazon S3.",
+  "description": "Storage service clients module of awesome-backup",
   "author": "Ryu Sato <ryu@weseek.co.jp>",
   "license": "MIT",
   "homepage": "https://github.com/ryu-sato/awesome-backup#readme",
@@ -12,35 +12,36 @@
   "bugs": {
     "url": "https://github.com/ryu-sato/awesome-backup/issues"
   },
+  "// Use Typescript as is, so Typescript must be enabled": "",
+  "main": "src/index.ts",
   "directories": {
     "test": "__tests__"
   },
   "files": [
     "dist"
   ],
-  "main": "dist/apps/mongodb-restore/src/bin/restore.js",
-  "bin": {
-    "restore": "bin/restore.js"
-  },
   "scripts": {
     "build": "tsc && tsc-alias",
-    "// for development": "",
-    "ts-node-dev": "cross-env NODE_ENV=development ts-node-dev -r tsconfig-paths/register",
-    "//// Cannot run in parallel due to shared S3, GCS mock": "",
-    "test": "cross-env NODE_ENV=test jest --testTimeout 10000 --passWithNoTests -i",
+    "// Since the risk of memory leaks increases when a few workers, we chose to use many workers even though the processing speed would be lower.": "",
+    "test": "cross-env NODE_ENV=test jest --maxWorkers 4 --ci --logHeapUsage",
     "lint": "eslint src --ext .ts",
     "lint:fix": "eslint src --ext .ts --fix"
   },
   "dependencies": {
-    "@awesome-backup/commands": "^0.0.1",
+    "@aws-sdk/client-s3": "^3.45.0",
+    "@google-cloud/storage": "^5.16.1",
+    "axios": "^0.24.0",
+    "axios-retry": "^3.2.4",
+    "cross-env": "^7.0.3",
     "bunyan": "^1.8.15",
     "universal-bunyan": "^0.9.2",
-    "browser-bunyan": "^1.6.3"
+    "browser-bunyan": "^1.6.3",
+    "date-fns": "^2.28.0",
+    "node-schedule": "^2.1.0",
+    "tmp": "^0.2.1"
   },
   "devDependencies": {
     "@awesome-backup/tsconfig": "^0.0.1",
-    "cross-env": "^7.0.3",
-    "tsconfig-paths": "^3.12.0",
     "@types/bunyan": "^1.8.8"
   },
   "engines": {
