+++ packages/storage-service-clients/src/gcs.ts
