+++ packages/storage-service-clients/src/index.ts
@@ -1,0 +1,6 @@
+export * from './interfaces';
+export * from './factory';
+export * from './s3';
+export * from './s3-config';
+export * from './gcs';
+export * from './types';
