+++ packages/storage-service-clients/src/interfaces/config/gcs.ts
