+++ packages/storage-service-clients/src/interfaces/config/s3.ts
