+++ packages/storage-service-clients/src/interfaces/index.ts
