+++ packages/storage-service-clients/src/interfaces/method-options/gcs.ts
