+++ packages/storage-service-clients/src/interfaces/method-options/s3.ts
