+++ packages/storage-service-clients/src/interfaces/uri/gcs.ts
