+++ packages/storage-service-clients/src/interfaces/uri/s3.ts
