+++ packages/storage-service-clients/tsconfig.json
@@ -1,0 +1,16 @@
+{
+  "extends": "@awesome-backup/tsconfig/tsconfig.base.json",
+  "compilerOptions": {
+    "outDir": "./dist",
+    "baseUrl": "."
+  },
+  "include": [
+    "src",
+  ],
+  "exclude": [
+    "node_modules",
+    "dist",
+    "src/public/**",
+    "../commands/src/commands/interfaces/coverage"
+  ]
+}
