+++ packages/tar/package.json
@@ -1,0 +1,40 @@
+{
+  "name": "@awesome-backup/tar",
+  "version": "0.0.1",
+  "description": "`tar` command wrapper",
+  "author": "Ryu Sato <ryu@weseek.co.jp>",
+  "license": "MIT",
+  "homepage": "https://github.com/ryu-sato/awesome-backup#readme",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ryu-sato/awesome-backup.git"
+  },
+  "bugs": {
+    "url": "https://github.com/ryu-sato/awesome-backup/issues"
+  },
+  "// Use Typescript as is, so Typescript must be enabled": "",
+  "main": "src/index.ts",
+  "directories": {
+    "test": "__tests__"
+  },
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "tsc && tsc-alias",
+    "// Since the risk of memory leaks increases when a few workers, we chose to use many workers even though the processing speed would be lower.": "",
+    "test": "cross-env NODE_ENV=test jest --maxWorkers 4 --ci --logHeapUsage",
+    "lint": "eslint src --ext .ts",
+    "lint:fix": "eslint src --ext .ts --fix"
+  },
+  "dependencies": {
+    "cross-env": "^7.0.3"
+  },
+  "devDependencies": {
+    "@awesome-backup/tsconfig": "^0.0.1"
+  },
+  "engines": {
+    "node": ">= 16.13.1 < 17",
+    "yarn": ">= 1.22.4"
+  }
+}
